Ignore "Last" tag in gen_category_ranges
`build_category_ranges` looks for continuous runs of code points with the same
category. For this to work there is not need to treat code points with the
"Last" tag in a special way: the only thing we are interested in are gapless
sequences of code points with the same category.

With the upgrade to Unicode 16.0.0, the previous code broke due to the
introduction of `KHITAN SMALL SCRIPT CHARACTER-18CFF`, which has the code point
`18CFF`. This code point has the category "Letter, Other". This codes begins a
new range as the previous code point was `18CD5`. The next code point, `<Tangut
Ideograph Supplement, First>` (`18D00`) also belongs to the same category as
does the final code point of the run `<Tangut Ideograph Supplement,
Last>` (`18D08`). When processing this final entry of the range, the code failed
the `charstr_ends_with(names[range->first], ", First>")` part of the assertion
as it assumed `<Tangut Ideograph Supplement, First>` would have started the
range.